+++ src/main/java/ru/vyarus/java/generics/resolver/GenericsResolver.java
@@ -1,7 +1,7 @@
 package ru.vyarus.java.generics.resolver;
 
 import ru.vyarus.java.generics.resolver.context.GenericsInfoFactory;
-import ru.vyarus.java.generics.resolver.context.TypeGenericsContext;
+import ru.vyarus.java.generics.resolver.context.GenericsContext;
 import ru.vyarus.java.generics.resolver.util.TypeUtils;
 
 /**
@@ -28,9 +28,9 @@ public final class GenericsResolver {
      *                      or to limit resolution depth)
      * @return resolved generics context object
      */
-    public static TypeGenericsContext resolve(final Class<?> type, final Class<?>... ignoreClasses) {
+    public static GenericsContext resolve(final Class<?> type, final Class<?>... ignoreClasses) {
         final Class<?> notPrimitiveType = TypeUtils.wrapPrimitive(type);
-        return new TypeGenericsContext(
+        return new GenericsContext(
                 GenericsInfoFactory.create(notPrimitiveType, ignoreClasses), notPrimitiveType);
     }
 }