+++ src/test/groovy/ru/vyarus/java/generics/resolver/BrokenHierarchyTest.groovy
@@ -1,7 +1,7 @@
 package ru.vyarus.java.generics.resolver
 
 import ru.vyarus.java.generics.resolver.context.AbstractGenericsContext
-import ru.vyarus.java.generics.resolver.context.TypeGenericsContext
+import ru.vyarus.java.generics.resolver.context.GenericsContext
 import ru.vyarus.java.generics.resolver.support.brokenhieararchy.BrokenHierarchyBase
 import ru.vyarus.java.generics.resolver.support.brokenhieararchy.BrokenHierarchyInterface
 import ru.vyarus.java.generics.resolver.support.brokenhieararchy.BrokenHierarchyRoot
@@ -19,7 +19,7 @@ class BrokenHierarchyTest extends Specification {
     def "Check broken hierarchy resolution"() {
 
         when: "resolving class with no generics set"
-        TypeGenericsContext context = GenericsResolver.resolve(BrokenHierarchyRoot).type(BrokenHierarchyBase)
+        GenericsContext context = GenericsResolver.resolve(BrokenHierarchyRoot).type(BrokenHierarchyBase)
         then: "generics resolved just from generic bound"
         context.generic("T") == Callable
         context.generic("K") == Object