+++ src/test/groovy/ru/vyarus/java/generics/resolver/DependantRawTypesTest.groovy
@@ -1,6 +1,6 @@
 package ru.vyarus.java.generics.resolver
 
-import ru.vyarus.java.generics.resolver.context.TypeGenericsContext
+import ru.vyarus.java.generics.resolver.context.GenericsContext
 import spock.lang.Specification
 
 /**
@@ -12,7 +12,7 @@ class DependantRawTypesTest extends Specification {
     def "Check dependant raw types"() {
 
         when: "dependant root generics declared"
-        TypeGenericsContext context = GenericsResolver.resolve(SelfPairProperties)
+        GenericsContext context = GenericsResolver.resolve(SelfPairProperties)
 
         then: "correctly resolved"
         context.genericsMap() == ["F": Object, "S": Object]