+++ src/test/groovy/ru/vyarus/java/generics/resolver/ToStringTest.groovy
@@ -1,6 +1,6 @@
 package ru.vyarus.java.generics.resolver
 
-import ru.vyarus.java.generics.resolver.context.TypeGenericsContext
+import ru.vyarus.java.generics.resolver.context.GenericsContext
 import ru.vyarus.java.generics.resolver.context.container.ParameterizedTypeImpl
 import ru.vyarus.java.generics.resolver.error.UnknownGenericException
 import ru.vyarus.java.generics.resolver.support.Model
@@ -22,7 +22,7 @@ class ToStringTest extends Specification {
     def "Complex to string"() {
 
         when:
-        TypeGenericsContext context = GenericsResolver.resolve(TSRoot).type(TSBase)
+        GenericsContext context = GenericsResolver.resolve(TSRoot).type(TSBase)
         then:
         context.genericsAsString() == ["Model", "ArrayList<SType<Model, Model[]>>"]
 
@@ -42,7 +42,7 @@ class ToStringTest extends Specification {
     def "Complex to string 2"() {
 
         when: "resolving all types of interface generics"
-        TypeGenericsContext context = GenericsResolver.resolve(Base).type(GenerifiedInterface)
+        GenericsContext context = GenericsResolver.resolve(Base).type(GenerifiedInterface)
         then: "everything is ok"
         context.genericsAsString() == ['Integer', 'String[]', 'List<String>', 'List<Set<String>>']
     }
@@ -57,7 +57,7 @@ class ToStringTest extends Specification {
     def "Wildcards to string"() {
 
         when: "to string wildcard"
-        TypeGenericsContext context = GenericsResolver.resolve(WCRoot).type(WCBase)
+        GenericsContext context = GenericsResolver.resolve(WCRoot).type(WCBase)
         then: "correct"
         context.genericAsString(0) == "Model"
         context.genericAsString(1) == "? super Model"
@@ -68,8 +68,8 @@ class ToStringTest extends Specification {
     def "Inner context to string"() {
 
         when: "reasolve type with outer generics"
-        TypeGenericsContext context = GenericsResolver.resolve(InnerTypesTest.Root)
-        TypeGenericsContext innerContext = context.fieldType(InnerTypesTest.Root.getDeclaredField('target'))
+        GenericsContext context = GenericsResolver.resolve(InnerTypesTest.Root)
+        GenericsContext innerContext = context.fieldType(InnerTypesTest.Root.getDeclaredField('target'))
 
         then: "to string properly selects type generics only"
         innerContext.toStringCurrentClass() == "Inner"